GameControllerSetUp
Made the GameController class, and placed it in a new MainGame scene. GameController instantiates the players up to player count in the saved spawn positions, and gives them the correct player number for controlling. SetUp could be called again later to reset positions of things. (Still needs to remove chicken, and zombies, assuming we don't want to reload the scene entirely when the player loses.)